Group_14/Source_Code/carbon_backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
deprecation==2.1.0
dj-database-url==2.3.0
Django==5.2
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==4.1
django-htmx==1.23.0
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
frozenlist==1.5.0
geographiclib==2.1
geopy==2.4.1
googlemaps==4.10.0
gotrue==2.11.4
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.8
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
joblib==1.5.3
matplotlib>=3.7.0
multidict==6.4.3
numpy==2.4.1
openai>=1.0.0
openpyxl==3.1.2
packaging==24.2
pandas==2.3.3
pillow>=11.0.0
plotly>=5.15.0
pluggy==1.5.0
postgrest==1.0.1
propcache==0.3.1
psycopg2-binary==2.9.10
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-dotenv>=1.0.0
pytz==2025.2
realtime==2.4.2
requests==2.31.0
scipy>=1.11.0
scikit-learn>=1.3.0
seaborn>=0.12.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.0
supafunc==0.9.4
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.3
urllib3==2.4.0
websockets==14.2
whitenoise==6.9.0
yarl==1.19.0
